MIS-Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
pyodbc==4.0.39
cryptography==41.0.4
bcrypt==4.0.1
Werkzeug==2.3.7 